+++ objectified-ui/public/WHATS_NEW.md
@@ -40,7 +40,8 @@ We continue to improve the platform based on your feedback with improvements and
     - Required
     - Default Value
     - Description
-  - Configure required OAuth2/OpenID Connect scopes per operation (#421)
+  - Configure required OAuth2/OpenID Connect scopes per operation
+  - Configure serialization style for parameterized variables: form, spaceDelimited, pipeDelimited, deepObject
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions